src/requirements.txt
fpdf==1.7.2
selenium==3.8.0
Pillow
